requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.3

# AI/ML dependencies
sentence-transformers==2.2.2
torch==2.1.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# API integrations
jira==3.5.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
celery==5.3.4
redis==5.0.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2

# API Documentation
flasgger==0.9.7.1

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.2

# Development tools
pre-commit==3.6.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.20.0

# Database migrations
alembic==1.13.1
pytest-cov==4.1.0
